% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbots.correcttion.plot.R
\name{abbots.correction.plot}
\alias{abbots.correction.plot}
\title{Creates Abbot's Correction Plots and Tables for Toxicology}
\usage{
abbots.correction.plot(adjusted.data, control, write = T, format = "pdf",
  subfolder = "Abbots_Correction", interval = 0.95)
}
\arguments{
\item{adjusted.data}{A data frame or data table. Must include control dose, 0, for each insecticide. This is input for this value is usually an ouput from dmc.correct}

\item{control}{A string of your control genotype. e.g "Line14"}

\item{write}{If TRUE graphs will be produced in the specified format and .csv tables will be generated. If FALSE ouput will be localized to R. Defaults to TRUE}

\item{format}{what format do you want your graph images? Possible values "pdf", "png","tiff","jpeg"}
}
\value{
Creates Abbot's correction Plots and Tables for Toxicology
}
\description{
Creates Abbot's Correction Plots and Tables for Toxicology
}
