% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc.control.add.R
\name{dmc.control.add}
\alias{dmc.control.add}
\title{Adds Controls for Each Pesticide You are Using}
\usage{
dmc.control.add(raw.data, key = NULL, new.key = F, solvent.overide = NA)
}
\arguments{
\item{raw.data}{Raw toxicology in standard Batterham Lab Format}

\item{key}{Do you want to specify your own insecticide:solvent key? If so put it in here. Default is NULL}

\item{new.key}{Do you want to specify your own insecticide solvent key? Defaults to FALSE}

\item{solvent.override}{You can manually set all of the solvents to "water" "dmso" or "ethanol" if you please. Defaults to NA}
}
\value{
returns data frame which has added the 0 dose for all of your insecticides
}
\description{
Adds Controls for Each Pesticide You are Using
}
