% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose.determine.R
\name{dose.determine}
\alias{dose.determine}
\title{Determines Doses for an Lc50 screen}
\usage{
dose.determine(min, max, n)
}
\arguments{
\item{min}{The minimum concentration that will be used in your screen. Ideally only a small number of individuals should die at this dose}

\item{max}{The minimum concentration that will be used in your screen. Ideally only a small number of individuals should survive at this dose}

\item{n}{The number of doses that you want to use in your screen. More is always better but is always more}
}
\value{
This function determines the doses to use for an Lc50 screen
}
\description{
Determines Doses for an Lc50 screen
}
