% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld.summary.R
\name{ld.summary}
\alias{ld.summary}
\title{Summarizes data into LC50 table}
\usage{
ld.summary(adjusted.data, conf.level = 0.95, ld.level = 50)
}
\arguments{
\item{adjusted.data}{data frame of toxicology in standard format}

\item{conf.level}{What do you want your confidence intervals to be % wise? defaults to .95}

\item{ld.level}{What LD level do you want to predict? defaults to 50 to calculate Ld50}
}
\value{
Calculates LD values for genotype:insecticide combination
}
\description{
Summarizes data into LC50 table
}
