% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.merge.R
\name{wiggle.merge}
\alias{wiggle.merge}
\title{Merge Wiggle Files}
\usage{
wiggle.merge(wd = ".", output.name = "merged.wiggle.data.csv", write = T)
}
\arguments{
\item{wd}{The working directory that your collection of .xls wiggle outputs are in}

\item{output.name}{A string of the desired name of the output file}

\item{write}{If TRUE the results will be written to a .csv file. Defaults to FALSE}
}
\value{
Merges wiggle index outputs from individual wells into one file
}
\description{
Merge Wiggle Files
}
