% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.parse.R
\name{wiggle.parse}
\alias{wiggle.parse}
\title{Parse Raw Wiggle Data}
\usage{
wiggle.parse(data, ouptut.name = "parsed.wiggle.data.csv", write = T)
}
\arguments{
\item{data}{data frame containing the raw wiggle data in two columns: "Image.Name" and "Wiggle.Index"}

\item{write}{If TRUE the results will be written to a .csv file. Defaults to FALSE}

\item{output.name}{A string of the desired name of the output file}
}
\value{
Parses wiggle index outputs into cleaner more readable data frame
}
\description{
Parse Raw Wiggle Data
}
