% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.plot.R
\name{wiggle.plot}
\alias{wiggle.plot}
\title{Wiggle Plot}
\usage{
wiggle.plot(rmr.data, control, other.genotypes = "all",
  colours = c("grey50", "red", "blue", "gold4", "purple", "orange"),
  write = T, format = "pdf", font = "serif")
}
\arguments{
\item{rmr.data}{A data frame containing wiggle index RMR values. Usually output of wiggle.rmr funciton}

\item{control}{A string of your control genotype. e.g "Line14"}

\item{write}{If TRUE graphs will be produced in the specified format and .csv tables will be generated. If FALSE ouput will be localized to R. Defaults to TRUE}

\item{format}{what format do you want your graph images? Possible values "pdf", "png","tiff","jpeg"}

\item{font}{Which font do you want your graph text to be in}
}
\value{
Creates standard Wiggle Index plots
}
\description{
Wiggle Plot
}
