% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiggle.rmr.R
\name{wiggle.rmr}
\alias{wiggle.rmr}
\title{Correct Wiggle Data}
\usage{
wiggle.rmr(data, file.name = "rmr.wiggle.data.csv", write = T)
}
\arguments{
\item{data}{parsed wiggle data frame containing the columns genotype, time, dose, well, rep, and date. Usually this data frame is the ouput of wiggle.parse}

\item{write}{If TRUE the results will be written to a .csv file. Defaults to FALSE}

\item{output.name}{A string of the desired name of the output file}
}
\value{
Corrects wiggle data by adding RMR value. See Denecke et. al 2017
}
\description{
Correct Wiggle Data
}
